requirements.txt
Flask
flask-restx
flask-bcrypt
flask-jwt-extended
flask-pymongo
pymongo
python-dotenv